+++ CHANGELOG.md
@@ -26,8 +26,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Refactored/Updated SingleLevelUsageAsBuilt maven module to support multi version.
 - Refactored/Updated PCF maven module to support multi version.
 - Added new flyway files.
-- Supporting new submodel Singlelevelbomasplanned
-- Support EDC 7
+- Supporting new submodel Singlelevelbomasplanned.
+- Support EDC 7.
+- Added new files for digital twin access rule support.
 
 ### Fixed
 - Remove garbage character from 'edc_request_template' path. Fixed [#147](https://github.com/eclipse-tractusx/managed-simple-data-exchanger-backend/issues/147).